Ajustes de retorno nos jogos favoritos e adicionando validações para exceção do Token JWT expirado

+++ game-service/src/main/java/com/TP/game_service/models/DTOs/BaseResponseDTO.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BaseResponseDTO<T> {
-    private Double count;
+    private int count;
     private String next;
     private String previous;
     private List<T> results;
